requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiogram-dialog==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
diagrams==0.23.4
Faker==25.2.0
frozenlist==1.4.1
graphviz==0.20.3
idna==3.7
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typed-ast==1.5.5
typing_extensions==4.11.0
validators==0.28.1
yarl==1.9.4
